Moved around mcmc samplers and epmgp
