+++ emukit/samplers/__init__.py
@@ -1,2 +1,5 @@
 # Copyright 2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 # SPDX-License-Identifier: Apache-2.0
+
+
+from .mcmc_sampler import AffineInvariantEnsembleSampler, McmcSampler  # noqa: F401
