+++ tests/emukit/bayesian_optimization/test_entropy_search.py
@@ -2,7 +2,7 @@ import numpy as np
 import pytest
 
 from emukit.bayesian_optimization.acquisitions import EntropySearch
-from emukit.bayesian_optimization.util.mcmc_sampler import AffineInvariantEnsembleSampler
+from emukit.samplers import AffineInvariantEnsembleSampler
 
 
 @pytest.fixture
